+++ src/lib/items/defaultItemRendere.js
@@ -10,26 +10,16 @@ export const defaultItemRenderer = ({
   const { left: leftResizeProps, right: rightResizeProps } = getResizeProps()
   return (
     <div {...getItemProps(item.itemProps)}>
-      {itemContext.useResizeHandle && itemContext.showInnerContentsRender ? (
-        <div {...leftResizeProps} />
-      ) : (
-        ''
-      )}
+      {itemContext.useResizeHandle ? <div {...leftResizeProps} /> : ''}
 
-      {itemContext.showInnerContentsRender && (
-        <div
-          className="rct-item-content"
-          style={{ maxHeight: `${itemContext.dimensions.height}` }}
-        >
-          {itemContext.title}
-        </div>
-      )}
+      <div
+        className="rct-item-content"
+        style={{ maxHeight: `${itemContext.dimensions.height}` }}
+      >
+        {itemContext.title}
+      </div>
 
-      {itemContext.useResizeHandle && itemContext.showInnerContentsRender ? (
-        <div {...rightResizeProps} />
-      ) : (
-        ''
-      )}
+      {itemContext.useResizeHandle ? <div {...rightResizeProps} /> : ''}
     </div>
   )
 }