+++ CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres (more or less) to [Semantic Versioning](http://semver.o
 
 ## Unreleased
 
+* rework item renderer to render the whole item using render prop and prop getters
+
+### Breaking
+* add scroll ref
 
 ### 0.18.2
 
-### Breaking
-* rework item renderer to render the whole item using render prop and prop getters
+
 
 ### Fixed
 
@@ -76,7 +79,6 @@ from 'react-calendar-timeline'
 ### Added
 
 * pass canvasTimeStart/End via timelineContext to the itemRenderer prop
-
 
 ### 0.17.0